requirements.txt
selenium==4.5.0
python-decouple==3.6